feat(sdk): move `fetch_url` into Deep Agents
Keep the dcode import path as a lazy compatibility re-export.
